Source applications from redis
Create a app-loader.js script that runs before the application starts.
This fetches all the apps from the google sheet and adds any apps to the
db that had not been added previously. Now, the app server its
applications from the db instead of the google sheet. It is FAST.

Next up, looking and Google OAuth for grading submission.