Move config and service keys to environment variables
I had to separate about the private_key embedded in the jwt.keys.json
file Google provides, because if you want to store the keys are JSON,
you need to remove all the newlines, but you can't remove the newlines
on a private key. So I just pulled it out and serialized it separately.