Add locking mechanism that remembers the user
If a user that has been seen before requests the page again, they will
get the same application; otherwise they will get a new one. There is a
fair amount to document about how this works with redis data structure,
but for now suffice to say that it stores expire times in each of the
locks, and before each checkout request invalidates all the locks that
expired prior to the current time.

Resolves: #1