Update apps-loader to ignore existing apps in set
Previously, I removed an NX flag that would prevent new apps in the
keyspace, which I removed so that apps could be updated on load. It is
important, however, to have an NX flag on the apps in the *set*, because
that keeps track of how many times each one has been graded.