Replace redis integration with SQLite
The redis integration was a lot of fun to reason about, but it has two
principal drawbacks:

1. Doing locking properly is complicated
2. Operating over a network boundary (even a local one) is complicated

As it turns out, fixing the second problem reduces the need for the
complex locking mechanism, because of node's single-threaded runtime. If
the "fetch random app" and "lock app" transcations never make use of
async functions, then there's no chance that a different request will
complete the "fetch random app" function and resolve to the same app. So
a synchronous SQLite library can fetch and lock apps quickly with no
risk of race conditions.

SQLite is also far easier to standup, and more than meets the kind of
request load that this application sees. Wins all around.